add food to cart api made

+++ swiggato/src/main/java/com/example/swiggato/model/Cart.java
@@ -18,7 +18,7 @@ public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     int id;
-    int cartTotal;
+    double cartTotal;
 
     @OneToOne
     @JoinColumn
